Refactor parsing to lib files
